Replace react-md tabs with home grown
