+++ premiser-ui/src/App.scss
@@ -93,14 +93,6 @@ strong {
   position: relative;
 }
 
-// Tabs that appear to be contiguous with the toolbar.  But to avoid handling the navigate ourselves, they are actually not in the toolbar
-.toolbarTabs {
-  background-color: $primary-color;
-  * {
-    color: $light-text-color;
-  }
-}
-
 .app-nav-drawer-header {
   @include center-contents;
 }
